backend/requirements.txt
setuptools>=68.0.0
flake8==7.2.0
flake8-isort==6.1.1
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==5.0.6
django-extensions==4.1
django-filter==24.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields>=3.7.0
idna==3.10
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
urllib3==2.4.0
black==24.3.0
isort==5.13.2
